+++ lib/_test_page.dart
@@ -1,9 +1,18 @@
 // ignore_for_file: unused_element,unused_import
+import 'dart:typed_data';
+
 import 'package:chaldea/components/components.dart';
 import 'package:path/path.dart' as p;
 
+import 'components/img_util.dart';
+
 void testFunction() async {
-  if (Platform.isMacOS) _reloadDebugDataset();
+  // if (Platform.isMacOS) _reloadDebugDataset();
+  Uint8List? img =
+      await db.runtimeData.screenshotController!.capture(pixelRatio: 2);
+  File(join(db.paths.tempDir, 'test1.png')).writeAsBytesSync(img!);
+  img = compressToJpg(src: img);
+  File(join(db.paths.tempDir, 'test2.jpg')).writeAsBytesSync(img);
   return;
 }